backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic-settings==2.10.1
python-multipart==0.0.20
pypdf==5.9.0
openai==1.106.1
PyJWT==2.10.1
